requirements.txt
discord
tortoise-orm
psycopg
python-dotenv
requests
asyncpg
aiohttp_cors
humanfriendly
wavelink
easy-pil
